restructured the dataset folder
there is now a dedicated file for fetching, and other utilities
