+++ dirty_cat/test/test_datasets.py
@@ -6,8 +6,8 @@ Test the datasets module
 import os
 from tempfile import mkstemp
 
-import dirty_cat.datasets as datasets
-import dirty_cat.utils as utils
+import dirty_cat.datasets.utils as datasets_utils
+import dirty_cat.test.utils as utils
 
 currdir = os.path.dirname(os.path.abspath(__file__))
 datadir = os.path.join(currdir, 'data')
@@ -16,7 +16,7 @@ url_request = None
 file_mock = None
 
 
-def setup_mock(true_module=datasets, mock_module=utils):
+def setup_mock(true_module=datasets_utils, mock_module=utils):
     global original_url_request
     global mock_url_request
     mock_url_request = mock_module.mock_request_get
@@ -24,7 +24,7 @@ def setup_mock(true_module=datasets, mock_module=utils):
     true_module.request_get = mock_url_request
 
 
-def teardown_mock(true_module=datasets):
+def teardown_mock(true_module=datasets_utils):
     global original_url_request
     true_module.request_get = original_url_request
 
@@ -32,10 +32,10 @@ def teardown_mock(true_module=datasets):
 @utils.with_setup(setup=setup_mock, teardown=teardown_mock)
 def test_fetch_file_overwrite():
     # overwrite non-exiting file.
-    test_dir = datasets.get_data_dir(name='test')
-    fil = datasets._fetch_file(url='http://foo/', data_dir=test_dir,
-                               overwrite=True
-                               , uncompress=False)
+    test_dir = datasets_utils.get_data_dir(name='test')
+    from dirty_cat.datasets import fetching
+    fil = fetching._fetch_file(url='http://foo/', data_dir=test_dir,
+                               overwrite=True, uncompress=False)
 
     assert os.path.exists(fil)
     with open(fil, 'r') as fp:
@@ -46,7 +46,7 @@ def test_fetch_file_overwrite():
         fp.write('some content')
 
     # Don't overwrite existing file.
-    fil = datasets._fetch_file(url='http://foo/', data_dir=test_dir,
+    fil = fetching._fetch_file(url='http://foo/', data_dir=test_dir,
                                overwrite=False, uncompress=False)
     assert os.path.exists(fil)
     with open(fil, 'r') as fp:
@@ -54,7 +54,7 @@ def test_fetch_file_overwrite():
 
     # Overwrite existing file.
     # Overwrite existing file.
-    fil = datasets._fetch_file(url='http://foo/', data_dir=test_dir,
+    fil = fetching._fetch_file(url='http://foo/', data_dir=test_dir,
                                overwrite=True, uncompress=False)
     assert os.path.exists(fil)
     with open(fil, 'r') as fp:
@@ -66,7 +66,7 @@ def test_md5_sum_file():
     out, f = mkstemp()
     os.write(out, b'abcfeg')
     os.close(out)
-    assert datasets._md5_sum_file(f) == '18f32295c556b2a1a3a8e68fe1ad40f7'
+    assert datasets_utils._md5_sum_file(f) == '18f32295c556b2a1a3a8e68fe1ad40f7'
     os.remove(f)
 
 # if __name__ == '__main__':
